requirements.txt
DateTime==5.3
pygame==2.5.2
Pillow==9.4.0